requirement.txt
streamlit==1.32.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-community==0.0.24
langchain-google-genai==0.0.6
chromadb==0.4.22
sentence-transformers==2.5.1
PyPDF2==3.0.1
SpeechRecognition==3.10.1
google-generativeai==0.3.2
edge-tts==6.1.9
asyncio==3.4.3
pygame==2.5.2 
pypdf
playsound==1.2.2
gtts
pyttsx3
pyaudio==0.2.14